Move shakespeare specific stuff to Yesod.Shakespeare
Widgets should have an interface for templates

The dependency chain is now

Yesod.Core.* -> Yesod.Widget -> Yesod.Shakespeare
