+++ yesod-core/yesod-core.cabal
@@ -30,7 +30,6 @@ library
                    , text                  >= 0.7
                    , template-haskell
                    , path-pieces           >= 0.1.2    && < 0.3
-                   , shakespeare           >= 2.0
                    , blaze-builder         >= 0.2.1.4  && < 0.5
                    , transformers          >= 0.2.2
                    , mtl
@@ -69,6 +68,7 @@ library
                    , word8
                    , auto-update
                    , semigroups
+                   , shakespeare
 
     exposed-modules: Yesod.Core
                      Yesod.Core.Content
@@ -80,6 +80,7 @@ library
                      Yesod.Core.Types
                      Yesod.Core.Unsafe
                      Yesod.Routes.TH.Types
+                     Yesod.Shakespeare
     other-modules:   Yesod.Core.Internal.Session
                      Yesod.Core.Internal.Request
                      Yesod.Core.Class.Handler
